refactor: update OID fields from Int to Long across bridge-api-models
Changed the data type for `oid` (Object ID) and `attackOid` from `Int` to `Long` in all model definitions and interfaces to support 64-bit identifiers.

Key changes:
- Updated `IL2Object` and `IL2Live` interfaces
- Updated all entity implementations: `L2User`, `L2Npc`, `L2Char`, `L2Live`, `L2Pet`, `L2Spawn`, `L2Object`
- Updated item-related models: `L2Item`, `L2ShopItem`, `L2MailItem`, `L2AucItem`, `L2Drop`
- Updated skill and message models: `L2Skill`, `L2Buff`, `LearnItem`, `L2ChatMessage`

+++ bridge-api-models/src/commonMain/kotlin/com/l2bot/bridge/models/L2Object.kt
@@ -18,7 +18,7 @@ data class L2Object(
     override val id: Int = 0,
 
     /** Unique object ID (OID) | Уникальный идентификатор объекта */
-    override val oid: Int = 0,
+    override val oid: Long = 0,
 
     /** Object validity | Валидность объекта */
     override val valid: Boolean = false,
