+++ bridge-api-models/src/commonMain/kotlin/com/l2bot/bridge/models/entities/L2User.kt
@@ -17,7 +17,7 @@ data class L2User(
     // IL2Object fields
     override val name: String = "",
     override val id: Int = 0,
-    override val oid: Int = 0,
+    override val oid: Long = 0,
     override val valid: Boolean = false,
     @SerialName("l2_class")
     override val l2Class: L2Class = L2Class.UNKNOWN,
@@ -79,7 +79,7 @@ data class L2User(
     override val abnormals: List<L2Buff> = emptyList(),
     override val equips: List<L2Item> = emptyList(),
     @SerialName("atk_oid")
-    override val attackOid: Int,
+    override val attackOid: Long,
 
     // IL2Char fields
     override val cp: Long = 0,
